refactor: move ad init code to utils folder
