+++ packages/ad/__tests__/android/ad.android.test.js
@@ -4,7 +4,7 @@ jest.mock(
   "../../src/utils/webview-event-callback-setup",
   () => "mockErrorHandler"
 );
-jest.mock("../../src/ad-init", () => () => "mockAdInit");
+jest.mock("../../src/utils/ad-init", () => () => "mockAdInit");
 jest.mock("WebView", () => "WebView");
 jest.mock("@times-components/watermark", () => "WaterMark");
