+++ packages/ad/__tests__/dom-context.native.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { Linking } from "react-native";
 import renderer from "react-test-renderer";
-
 import DOMContextNative from "../src/dom-context";
 
 // prevent function sources appearing in snapshots
@@ -9,7 +8,7 @@ jest.mock(
   "../src/utils/webview-event-callback-setup",
   () => "mockErrorHandler"
 );
-
+jest.mock("../src/utils/ad-init", () => {});
 jest.mock("WebView", () => "WebView"); // https://github.com/facebook/react-native/issues/12440
 
 export default () => {
@@ -69,20 +68,6 @@ export default () => {
         }
       });
     expect(f).not.toThrowError();
-  });
-
-  it("renders", () => {
-    const component = renderer.create(
-      <DOMContextNative
-        height={200}
-        width={200}
-        init={() => {}}
-        data={{ foo: "bar" }}
-      />
-    );
-    const s = component.toJSON();
-
-    expect(s).toMatchSnapshot();
   });
 
   it("calls onRenderComplete when it receives a renderComplete event from the webview", () => {
