+++ packages/ad/src/dom-context-prop-types.js
@@ -3,7 +3,6 @@ import PropTypes from "prop-types";
 export const propTypes = {
   data: PropTypes.shape({}),
   height: PropTypes.number.isRequired,
-  init: PropTypes.func.isRequired,
   onRenderComplete: PropTypes.func,
   onRenderError: PropTypes.func,
   width: PropTypes.number.isRequired
