+++ packages/ad/src/dom-context.web.js
@@ -1,10 +1,11 @@
 /* eslint-env browser */
 import React, { Component } from "react";
+import init from "./utils/ad-init";
 import { propTypes, defaultProps } from "./dom-context-prop-types";
 
 class DOMContext extends Component {
   componentDidMount() {
-    const { init, data } = this.props;
+    const { data } = this.props;
 
     this.initExecuting = true;
     this.hasUnmounted = false;
